Add types to ingredientDetails
